+++ sql/expression/function/aggregation/common.go
@@ -117,88 +117,3 @@ func (a unaryAggBase) Description() string {
 func windowResolved(w *sql.WindowDefinition) bool {
 	return expression.ExpressionsResolved(append(w.OrderBy.ToExpressions(), w.PartitionBy...)...)
 }
-
-type naryAggBase struct {
-	expression.NaryExpression
-	window       *sql.WindowDefinition
-	functionName string
-	description  string
-	typ          sql.Type
-}
-
-func (a *naryAggBase) NewWindowFunction() (sql.WindowFunction, error) {
-	panic("unaryAggBase is a base type, type must implement NewWindowFunction")
-}
-
-func (a *naryAggBase) NewBuffer() (sql.AggregationBuffer, error) {
-	panic("unaryAggBase is a base type, type must implement NewWindowFunction")
-}
-
-// WithWindow returns a new unaryAggBase to be embedded in wrapping type
-func (a *naryAggBase) WithWindow(window *sql.WindowDefinition) (sql.Aggregation, error) {
-	na := *a
-	na.window = window
-	return &na, nil
-}
-
-func (a *naryAggBase) Window() *sql.WindowDefinition {
-	return a.window
-}
-
-func (a *naryAggBase) String() string {
-	return fmt.Sprintf("%s(%s)", a.functionName, a.ChildExpressions)
-}
-
-func (a *naryAggBase) Type() sql.Type {
-	return a.typ
-}
-
-func (a *naryAggBase) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
-	return nil, ErrEvalUnsupportedOnAggregation.New(a.FunctionName())
-}
-
-func (a *naryAggBase) Children() []sql.Expression {
-	children := a.ChildExpressions
-	if a.window != nil {
-		children = append(children, a.window.ToExpressions()...)
-	}
-	return children
-}
-
-func (a *naryAggBase) Resolved() bool {
-	if _, ok := a.ChildExpressions[0].(*expression.Star); ok {
-		return true
-	} else if !a.ChildExpressions[0].Resolved() {
-		return false
-	}
-	if a.window == nil {
-		return true
-	}
-	return windowResolved(a.window)
-}
-
-// WithChildren returns a new unaryAggBase to be embedded in wrapping type
-func (a *naryAggBase) WithChildren(children ...sql.Expression) (sql.Expression, error) {
-	if len(children) < 1 {
-		return nil, sql.ErrInvalidChildrenNumber.New(a, len(children), 1)
-	}
-
-	na := *a
-	na.NaryExpression = expression.NaryExpression{ChildExpressions: children}
-	if len(children) > 1 && a.window != nil {
-		w, err := a.window.FromExpressions(children[1:])
-		if err != nil {
-			return nil, err
-		}
-		return na.WithWindow(w)
-	}
-	return &na, nil
-}
-
-func (a *naryAggBase) FunctionName() string {
-	return a.functionName
-}
-
-func (a *naryAggBase) Description() string {
-	return a.description
-}
